fix(web): restore reliable Vercel builds
