+++ agent-docs/references/testing-ci-map.md
@@ -407,18 +407,21 @@ supported provider credential.
   Repeated exact-head Standard previews remain the external acceptance proof.
   The next exact-head Standard preview still OOM-killed Turbopack, so the
   catalog correction is retained as a boundary fix but was not sufficient
-  capacity proof on Next 16.2.6. Production and Linux CI now use Next 16.3's
-  default Turbopack path through the same shared production-build selector. The
-  Workflow integration runs through its native Next integration: exact-head CI
-  proves the complete compile, type-validation, static-generation, and
-  directive-discovery path, while focused Stripe and phone-call suites prove
-  the existing `workflow/api.start` wrappers and step contracts. Two
-  forced-cold exact-head Standard previews completed without OOM: compilation
-  took 91 and 87 seconds, TypeScript validation took 54 and 55 seconds, all 233
-  pages took 10.0 and 10.8 seconds, and each Vercel build stage completed in
-  four minutes. These repeated previews remain the external memory acceptance
-  proof. The accepted candidate preserves the heap split and all route/type
-  validation. The advisory budget is
+  capacity proof on Next 16.2.6. A later Next 16.3 Turbopack trial completed two
+  forced-cold exact-head Standard previews in about four minutes, but production
+  subsequently reproduced the intermittent failure: on 2026-08-14 two builds
+  remained in Turbopack compilation until Vercel's build-duration ceiling and a
+  third exited 137 with Vercel's explicit container-OOM report. Production and
+  Linux CI therefore use the supported Webpack fallback through the same shared
+  production-build selector. The Next config explicitly enables the isolated
+  Webpack build worker and memory optimizations because Workflow contributes
+  Webpack configuration. Three consecutive forced-cold Webpack previews, a
+  later integration preview, and the final corrected head previously completed
+  on the Standard builder without OOM. Exact-head CI proves the complete
+  compile, type-validation, static-generation, and directive-discovery path,
+  while focused Stripe and phone-call suites prove the existing
+  `workflow/api.start` wrappers and step contracts. The accepted correction
+  preserves the heap split and all route/type validation. The advisory budget is
   a cgroup-unit model of Vercel Standard's 8 GB build machine: 7.2 GB available
   to the build cgroup and a 0.8 GB reserve for OS/container overhead outside it
   at the ceiling. The legacy-named guard budget override must stay strictly
