+++ apps/web/app/font-files.ts
@@ -10,18 +10,17 @@ const bundledWebDir = join(dirname(fileURLToPath(import.meta.url)), "..");
  * `node:fs` at render time. Each candidate below covers a different way this
  * resolution has been observed to break, cheapest and most reliable first.
  *
- * 1. Module-relative. Turbopack compiles `import.meta.url` to a path resolved
- *    from the emitted chunk's own location at runtime, so this relocates with
- *    the function and does not depend on the working directory. Verified
- *    against the emitted bundle by `scripts/check-og-emitted-runtime.ts`.
- * 2. Working directory plus `apps/web`. The deployed serverless function's
+ * 1. Working directory plus `apps/web`. The deployed serverless function's
  *    filesystem is shaped like the monorepo root, so this is where the assets
- *    live if the bundler ever hands back a build-machine path that no longer
- *    exists at runtime (the reported production symptom, an ENOENT on an
- *    absolute build path). It is also the layout the emitted-runtime check
- *    materializes.
- * 3. Bare working directory. Local dev, tests, and `next build` all run from
+ *    live. Trying this first also keeps Webpack from touching the build-time
+ *    source path it inlines for `import.meta.url`. This is the layout the
+ *    emitted-runtime check materializes.
+ * 2. Bare working directory. Local dev, tests, and `next build` all run from
  *    `apps/web`.
+ * 3. Module-relative. Turbopack compiles `import.meta.url` to a path resolved
+ *    from the emitted chunk's own location at runtime, so this relocates with
+ *    the function. Webpack inlines the source location instead, which is why
+ *    this candidate must remain last.
  *
  * The files are traced into every OG/card function by
  * `outputFileTracingIncludes` in `next.config.ts` and guarded by
@@ -33,9 +32,9 @@ export function ogAssetCandidatePaths(
 ): string[] {
   const base = cwd ?? /* turbopackIgnore: true */ process.cwd();
   return [
-    join(bundledWebDir, relativePath),
     join(base, "apps/web", relativePath),
     join(base, relativePath),
+    join(bundledWebDir, relativePath),
   ];
 }
