+++ apps/web/scripts/run-production-next-build.sh
@@ -20,7 +20,7 @@ else
 fi
 
 next_bin="$(node -p 'require.resolve("next/dist/bin/next")')"
-printf '[apps/web build] Next memory policy: parent_old_space_mb=%s typecheck_worker_old_space_mb=%s\n' \
+printf '[apps/web build] Next memory policy: compiler=webpack parent_old_space_mb=%s typecheck_worker_old_space_mb=%s\n' \
   "$parent_old_space_mb" \
   "$typecheck_worker_old_space_mb"
-exec node "--max-old-space-size=$parent_old_space_mb" "$next_bin" build
+exec node "--max-old-space-size=$parent_old_space_mb" "$next_bin" build --webpack
